refactor: The files of BlockModifier have been moved to codemetropolis.toolchain.commons.blockmodifier to have them in a separate location. Additionally, errors resulting from package and import modifications have been fixed.
BREAKING CHANGE: The program is in a non-executable state!
