+++ sources/commons/src/main/java/codemetropolis/toolchain/commons/blockmodifier/ext/FileHandler.java
@@ -1,4 +1,4 @@
-package codemetropolis.toolchain.commons.ext;
+package codemetropolis.toolchain.commons.blockmodifier.ext;
 
 import java.io.File;
 import java.io.FileInputStream;
