+++ sources/commons/src/main/java/codemetropolis/toolchain/commons/blockmodifier/ext/NBTTag.java
@@ -1,4 +1,4 @@
-package codemetropolis.toolchain.commons.ext;
+package codemetropolis.toolchain.commons.blockmodifier.ext;
 
 import java.io.DataInputStream;
 import java.io.DataOutputStream;
